:truck: move explore settings to correct place
