+++ lib/services/navigation_service.dart
@@ -51,8 +51,8 @@ import 'package:Okuna/pages/home/pages/menu/pages/settings/pages/account_setting
 import 'package:Okuna/pages/home/pages/menu/pages/settings/pages/account_settings/pages/user_language_settings/user_language_settings.dart';
 import 'package:Okuna/pages/home/pages/menu/pages/settings/pages/application_settings.dart';
 import 'package:Okuna/pages/home/pages/menu/pages/settings/pages/developer_settings.dart';
-import 'package:Okuna/pages/home/pages/menu/pages/settings/pages/explore_settings/explore_settings.dart';
-import 'package:Okuna/pages/home/pages/menu/pages/settings/pages/explore_settings/widgets/top_posts_excluded_communities.dart';
+import 'package:Okuna/pages/home/pages/search/widgets/top_posts/pages/top_posts_settings/pages/top_posts_excluded_communities/top_posts_excluded_communities.dart';
+import 'package:Okuna/pages/home/pages/search/widgets/top_posts/pages/top_posts_settings/top_posts_settings.dart';
 import 'package:Okuna/pages/home/pages/menu/pages/settings/settings.dart';
 import 'package:Okuna/pages/home/pages/menu/pages/useful_links.dart';
 import 'package:Okuna/pages/home/pages/menu/pages/user_invites/pages/user_invite_detail.dart';
